+++ lib/panels/tree/display.dart
@@ -35,6 +35,8 @@ import 'package:rattle/providers/stdout.dart';
 import 'package:rattle/r/extract_tree.dart';
 import 'package:rattle/widgets/show_markdown_file.dart';
 
+import '../../widgets/pages.dart';
+
 /// The tree panel displays the tree instructions or the tree biuld output.
 
 class TreeDisplay extends ConsumerStatefulWidget {
@@ -45,106 +47,37 @@ class TreeDisplay extends ConsumerStatefulWidget {
 }
 
 class TreeDisplayState extends ConsumerState<TreeDisplay> {
-  late PageController _pageController;
-  int _currentPage = 0;
   // number of pages available
-  int numPages = 2;
+  // calculate from the pages number
+  // int numPages = 2;
 
-  void _goToPreviousPage() {
-    if (_currentPage > 0) {
-      _pageController.animateToPage(
-        _currentPage - 1,
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeInOut,
-      );
-    }
-  }
-
-  void _goToNextPage() {
-    if (_currentPage < numPages - 1) {
-      _pageController.animateToPage(
-        _currentPage + 1,
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeInOut,
-      );
-    }
-  }
-
-  void goToResultPage() {
-    debugPrint('go to result page');
-    // TODO yyx 20240624 might need change when we have more pages than 2.
-    _goToNextPage();
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _pageController = PageController(initialPage: 0);
-  }
 
   @override
   Widget build(BuildContext context) {
     String stdout = ref.watch(stdoutProvider);
     String content = rExtractTree(stdout);
 
-    return Row(
-      children: [
-        IconButton(
-          icon: Icon(
-            Icons.arrow_left,
-            color: _currentPage > 0 ? Colors.black : Colors.grey,
-            size: 32,
-          ),
-          onPressed: _currentPage > 0 ? _goToPreviousPage : null,
-        ),
-        Expanded(
-          child:
-              // To avoid this error, provide a height limit at a higher (parent) level using the expanded() widget.
-              // Horizontal viewport was given unbounded height.
-              // Viewports expand in the cross axis to fill their container and constrain their children to match
-              // their extent in the cross axis. In this case, a horizontal viewport was given an unlimited amount of
-              // vertical space in which to expand.
-
-              // The relevant error-causing widget was:
-              //   PageView PageView:file:///Users/yinyixiang/repo/my_rattleng/lib/panels/tree/display.dart:112:20
-              PageView(
-            controller: _pageController,
-            onPageChanged: (index) {
-              setState(() {
-                _currentPage = index;
-              });
-            },
+    return Pages(
             children: [
-              showMarkdownFile(treeIntroFile, context),
-              Container(
-                decoration: sunkenBoxDecoration,
-                width: double.infinity,
-                padding: const EdgeInsets.only(left: 10),
-                child: content.isEmpty
-                    ? const Center(
-                        child: Text(
-                          'Click the build button to see the result',
-                        ),
-                      )
-                    : SingleChildScrollView(
-                        child: SelectableText(
-                          content,
-                          style: monoTextStyle,
-                        ),
-                      ),
+    showMarkdownFile(treeIntroFile, context),
+    Container(
+      decoration: sunkenBoxDecoration,
+      width: double.infinity,
+      padding: const EdgeInsets.only(left: 10),
+      child: content.isEmpty
+          ? const Center(
+              child: Text(
+                'Click the build button to see the result',
               ),
+            )
+          : SingleChildScrollView(
+              child: SelectableText(
+                content,
+                style: monoTextStyle,
+              ),
+            ),
+    ),
             ],
-          ),
-        ),
-        IconButton(
-          icon: Icon(
-            Icons.arrow_right,
-            size: 32,
-            color: _currentPage < numPages - 1 ? Colors.black : Colors.grey,
-          ),
-          onPressed: _currentPage < numPages - 1 ? _goToNextPage : null,
-        ),
-      ],
-    );
+          );
   }
 }
